src/backend/requirements.txt
psycopg2-binary
google-genai
tqdm
pandas
numpy
pathlib
langchain-google-vertexai
flask
flask-cors
fastapi
uvicorn
gunicorn
google-cloud-storage
google-adk
google-cloud-aiplatform>=1.38.0
deprecated>=1.2.13
toolbox-core
PyJWT
